+++ package.json
@@ -13,7 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.0",
-    "express": "^4.17.3",
-    "nodemon": "^2.0.15"
+    "express": "^4.18.2",
+    "nodemon": "^2.0.22"
   }
 }
